Rewritten implementation to use reduce-side joins.
Implementing the exact algorithm that SLCT uses did not work well if
the word counts would exceed allowed memory usage on each node/map.

Instead, I wanted a more generic solution that would work also for
less common patterns. The solution was to use sorted reduce joins to map
the (truncated) word counts to each log row. The drawback of this
solution is obviously that it requires the nodes to pass over the data
more than 2 times (which is the case of the original SLCT algorithm
presented in their paper).

This commit has been verified to work on a small testcase. While there
are a couple of very specific tests, I intend to very soon add some
integration tests that verifies that things are behaving exactly the
way the should be.